fix: One component per file Breadcrumb (#449)
* Rename file

* Rename temp

* Restore original

* Rename to private file

* Separate item
