+++ src/index.js
@@ -4,8 +4,7 @@ export { default as Badge } from './Badge/Badge';
 export { default as Label } from './Badge/Label';
 export { default as Status } from './Badge/Status';
 export { default as Counter } from './Badge/Counter';
-export { Breadcrumb } from './Breadcrumb/Breadcrumb';
-export { BreadcrumbItem } from './Breadcrumb/Breadcrumb';
+export { default as Breadcrumb } from './Breadcrumb/Breadcrumb';
 export { default as Button } from './Button/Button';
 export { default as ButtonGroup } from './Button/ButtonGroup';
 export { Calendar } from './Calendar/Calendar';
